Fix: cache_src_id must hash with word-size-native UInt, not UInt64
Base.hash's two-argument fallback requires the seed to be a UInt
(word-size-native: UInt64 on 64-bit, UInt32 on 32-bit). cache_src_id
hardcoded UInt64(inc.hash) as that seed, so on 32-bit Julia the call
doesn't match any hash method and throws:

  MethodError: no method matching hash(::UInt64, ::UInt64)

This fires every time Revise's package-watching callback parses a
newly precompiled package's cache header (Julia 1.11+, via
Base.CacheHeaderIncludes), i.e. on nearly every `using` of a not-yet-
loaded package once Revise is active -- non-fatal, but a noisy @error
on every load.

inc.fsize::UInt64 and inc.hash::UInt32 are genuinely fixed-width
fields from Julia's own cache-header format and are unchanged; only
the conversion target for hash's seed argument moves from UInt64 to
UInt.

Confirmed on Julia 1.12.7 (i686-w64-mingw32, via juliaup add
1.12.7~x86): `using Revise` followed by `using <any not-yet-loaded
package>` previously threw the MethodError above on every load; with
this fix it completes cleanly. No change in behavior on 64-bit
(UInt === UInt64 there).

+++ src/pkgs.jl
@@ -135,7 +135,7 @@ end
 # Julia 1.11+ records a size and a content hash; 1.10 records only the source file's
 # mtime, so on that version a file merely touched between two builds reads as changed.
 @static if VERSION >= v"1.11.0-DEV.683"    # https://github.com/JuliaLang/julia/pull/49866
-    cache_src_id(inc) = hash(inc.fsize, UInt64(inc.hash))
+    cache_src_id(inc) = hash(inc.fsize, UInt(inc.hash))
 else
     cache_src_id(inc) = hash(inc.mtime)
 end
